Move template* packages to `internal/helm`…
… they are only used by `helm` so let's *hide* them here.
