+++ internal/helm/helm.go
@@ -11,13 +11,13 @@ import (
 	yaml "gopkg.in/yaml.v2"
 
 	"github.com/docker/app/internal"
+	"github.com/docker/app/internal/helm/templateconversion"
+	"github.com/docker/app/internal/helm/templateloader"
+	"github.com/docker/app/internal/helm/templatev1beta2"
 	"github.com/docker/app/internal/packager"
 	"github.com/docker/app/internal/render"
 	"github.com/docker/app/internal/settings"
 	"github.com/docker/app/internal/slices"
-	"github.com/docker/app/internal/templateconversion"
-	"github.com/docker/app/internal/templateloader"
-	"github.com/docker/app/internal/templatev1beta2"
 	"github.com/docker/app/internal/types"
 	"github.com/docker/cli/cli/command/stack/kubernetes"
 	"github.com/docker/cli/cli/compose/loader"
